+++ signal_bot/__main__.py
@@ -1,3 +1,8 @@
-from signal_bot.main import run
+import sys
+from signal_bot.main import run, run_cli
+from signal_bot.config import load_config
 
-run()
+if "--cli" in sys.argv:
+    run_cli(load_config())
+else:
+    run()